Feature ETP-2450: Refactor PartitionedConstraintsHandling in utilities classes.
